upgrade to kepler@2.1.0-alpha.1
